Inheritance, DynamicCast and too much time
I have a working prototype of how to get the Sequence numbers that will likely time out. Amazingly, I was able to pull it off after a couple of changes by using DynamicCast after obtaining the installed Application via the Node::GetApplication function and converting to my type PriConsumer.

This seemed to enable everything else to work.

I am still confused as to what exactly changed, but I am glad I was able to inherit a class, extend it, read protected values and then use them in the main code.